requirements.txt
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
django-cors-headers==4.9.0
django-pwa==2.0.1
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pillow==12.1.1
python-dotenv==1.0.1
python-telegram-bot==22.6
qrcode==8.2
requests==2.32.5
sqlparse==0.5.5
stripe==14.4.1
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
